requirements.txt
# path: requirements.txt
# Telegram AI Group Finance Assistant - Dependencies

# Telegram Bot
python-telegram-bot[ext]==20.7

# Google APIs
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP Client (for AI API calls)
aiohttp==3.9.1
httpx==0.25.2

# Async utilities
asyncio-throttle==1.0.2

# Environment and Configuration
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.2

# Date/time utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
colorlog==6.8.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.7.1
flake8==6.1.0

# Type stubs
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
